fixed deletion of learning modules
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@3960 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ content/classes/class.ilLMObject.php
@@ -331,7 +331,10 @@ class ilLMObject
 		while($obj_rec = $obj_set->fetchRow(DB_FETCHMODE_ASSOC))
 		{
 			$lm_obj =& ilLMObjectFactory::getInstance($a_cobj, $obj_rec["obj_id"]);
-			$lm_obj->delete();
+			if (is_object($lm_obj))
+			{
+				$lm_obj->delete();
+			}
 		}
 	}
